Make "mailer" service optional
