+++ Mailer/TwigSwiftMailer.php
@@ -11,34 +11,15 @@
 
 namespace FOS\UserBundle\Mailer;
 
-use FOS\UserBundle\Model\UserInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 /**
  * @author Christophe Coevoet <stof@notk.org>
+ *
+ * @deprecated TwigSwiftMailer class is deprecated since 2.1, use TwigMailer instead.
  */
-class TwigSwiftMailer implements MailerInterface
+class TwigSwiftMailer extends TwigMailer
 {
-    /**
-     * @var \Swift_Mailer
-     */
-    protected $mailer;
-
-    /**
-     * @var UrlGeneratorInterface
-     */
-    protected $router;
-
-    /**
-     * @var \Twig_Environment
-     */
-    protected $twig;
-
-    /**
-     * @var array
-     */
-    protected $parameters;
-
     /**
      * TwigSwiftMailer constructor.
      *
@@ -49,74 +30,10 @@ class TwigSwiftMailer implements MailerInterface
      */
     public function __construct(\Swift_Mailer $mailer, UrlGeneratorInterface $router, \Twig_Environment $twig, array $parameters)
     {
-        $this->mailer = $mailer;
-        $this->router = $router;
-        $this->twig = $twig;
-        $this->parameters = $parameters;
-    }
+        @trigger_error(sprintf(
+            '%s is deprecated, use %s instead.', __CLASS__, TwigMailer::class
+        ), E_USER_DEPRECATED);
 
-    /**
-     * {@inheritdoc}
-     */
-    public function sendConfirmationEmailMessage(UserInterface $user)
-    {
-        $template = $this->parameters['template']['confirmation'];
-        $url = $this->router->generate('fos_user_registration_confirm', array('token' => $user->getConfirmationToken()), UrlGeneratorInterface::ABSOLUTE_URL);
-
-        $context = array(
-            'user' => $user,
-            'confirmationUrl' => $url,
-        );
-
-        $this->sendMessage($template, $context, $this->parameters['from_email']['confirmation'], (string) $user->getEmail());
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function sendResettingEmailMessage(UserInterface $user)
-    {
-        $template = $this->parameters['template']['resetting'];
-        $url = $this->router->generate('fos_user_resetting_reset', array('token' => $user->getConfirmationToken()), UrlGeneratorInterface::ABSOLUTE_URL);
-
-        $context = array(
-            'user' => $user,
-            'confirmationUrl' => $url,
-        );
-
-        $this->sendMessage($template, $context, $this->parameters['from_email']['resetting'], (string) $user->getEmail());
-    }
-
-    /**
-     * @param string $templateName
-     * @param array  $context
-     * @param array  $fromEmail
-     * @param string $toEmail
-     */
-    protected function sendMessage($templateName, $context, $fromEmail, $toEmail)
-    {
-        $template = $this->twig->load($templateName);
-        $subject = $template->renderBlock('subject', $context);
-        $textBody = $template->renderBlock('body_text', $context);
-
-        $htmlBody = '';
-
-        if ($template->hasBlock('body_html', $context)) {
-            $htmlBody = $template->renderBlock('body_html', $context);
-        }
-
-        $message = (new \Swift_Message())
-            ->setSubject($subject)
-            ->setFrom($fromEmail)
-            ->setTo($toEmail);
-
-        if (!empty($htmlBody)) {
-            $message->setBody($htmlBody, 'text/html')
-                ->addPart($textBody, 'text/plain');
-        } else {
-            $message->setBody($textBody);
-        }
-
-        $this->mailer->send($message);
+        parent::__construct($mailer, $router, $twig, $parameters);
     }
 }
