+++ Resources/config/mailer.xml
@@ -17,7 +17,7 @@
 
     <services>
         <service id="fos_user.mailer.default" class="FOS\UserBundle\Mailer\Mailer" public="false">
-            <argument type="service" id="mailer" />
+            <argument type="service" id="mailer" on-invalid="null" />
             <argument type="service" id="router" />
             <argument type="service" id="templating" />
             <argument type="collection">
@@ -31,8 +31,8 @@
             <tag name="fos_user.requires_swift" />
         </service>
 
-        <service id="fos_user.mailer.twig_swift" class="FOS\UserBundle\Mailer\TwigSwiftMailer" public="false">
-            <argument type="service" id="mailer" />
+        <service id="fos_user.mailer.twig_swift" class="FOS\UserBundle\Mailer\TwigMailer" public="false">
+            <argument type="service" id="mailer" on-invalid="null" />
             <argument type="service" id="router" />
             <argument type="service" id="twig" />
             <argument type="collection">
